+++ apyfal/configuration.py
@@ -51,6 +51,9 @@ METERING_TMP = '/tmp/meteringServer'
 #: Metering Client configuration
 METERING_CLIENT_CONFIG = '/etc/sysconfig/meteringclient'
 
+#: Apyfal directory in user home
+APYFAL_HOME = _os_path.expanduser('~/.apyfal/')
+
 __all__ = ['create_configuration', 'Configuration',
            'accelerator_executable_available',
            'ACCELERATOR_EXECUTABLE', 'ACCELERATOR_TMP_ROOT',